backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
mysqlclient==2.2.4
python-dotenv==1.0.1
langchain==0.1.9
openai==1.12.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2
pyjwt==2.8.0